Add files via upload
The objective of this project is to build a machine learning model that can predict whether a
borrower will default on a loan or not. The project will use the LendingClub dataset, which is
available on Kaggle. The LendingClub is a peer-to-peer lending platform that connects
borrowers with investors. The dataset contains information on loans issued by LendingClub
between 2007 and 2018, including the borrower’s credit score, employment history, loan
amount, interest rate, and loan status.